Fix bug in H&T - only generate MerchantInfo when we have all the data
We had a bug where we'd attempt a setActiveMerchant with a merchant object
which did not have an access_token.  Here I change our code which generates
a Merchant object from JSON to only return a merchant object if we have an
access_token in the JSON.  This causes the H&T login code to not hit an assert.

Instead we now return to kicking off the PPSNeedPayPalSetupViewController step
which takes the user on the steps needed to log into paypal and finally receive
an access token.