Update to latest SDK build plus updates to the SDKSampleApp
+ Update to the latest SDK.  Contains bug fixes, enhanced CAL logging,
and a refinement of the signature related API in PPHTransacitonManager.
When processPayment completes the PPHTransactionRecord that's returned
now contains the isSignatureRequired flag.  You no longer need to catch/read
this flag in the onPostAuthorize callout.

+ SDKSampleApp now shows how to install a logging delegate and monitor
the logs of the SDK

+ Fix some UI bugs in the Login screen.  (Broken spinner, ability to hit the
login button multiple times)

+ SDKSampleApp now allows you to supply a signature and finalize the transaction
when required.

+ SDKSampleApp now has a settings screen where we can access information about
the attached hardware reader.

+ The transaction screen now no longer uses the CardReaderManager.  It was only
in there so we could provide information about the currently attached hardware
reader. That was removed from the transaction VC to simplify the amount of code
in the transaction VC.

+ For cash payments set the PPHTransactionManager into a state where it won't
power up / scan for hardware readers when simply recording the cash payment.

+ When a transaction that needed a signature is complete let's pop to the root
so we can start another transaction.