Fix Live logins.   Enhance setBaseAPIURL documentation.
+ We had a nasty bug where we were setting the base URI to the Live
URL when we wanted Live.  This causes some internal code in the SDK to believe
we're NOT LIVE.  It turns out that when we want to run against Live we must
set the baseAPIURL to nil.   Here I modify the SDKSampleApp to demonstrate
this to partners.

+ Greatly expand the documentation in STAppDelegate that describes how to use
[PayPalHereSDK setBaseAPIURL].   I now give explicit examples of how to set it
for Live, Sandbox, and for a stage.