Fix a bug in our SDKSampleApp's login code.
If you selected a new stage (say stage mb006), when you returned to
the login page we'd successfully display 'Pointed To: stage2mb006'.  However,
when we build the login request (which happens when you tap the login
button) we'll place the old stage (say stage2mb023) into the httpBody
of the request for the servername param.  Then we were fetching the correct
server name and using it in the GET call.  So, the request was mismatched.

+ Fix sequence of events in the loginButtonPressed handler so that we
don't end up with the request having out of synch stage server names.

+ Don't load the stage server we are going to use from a UI control.  Instead
fetch it from our model which in this case is just the app delegate holding some
values.

+ Move some nsstring properties from strong to copy