Fix bug in EMVSampleApp UI where you can't select a stage
The bug:
 1) tap in the username field
 2) enter some characters
 3) tap the 'stage' button
 4) action sheet appears
 5) attempt to pick the item in the bottom of the list
Expected:
  Item to become selected and the action sheet to dismiss
Actual:
  You can't select any item in the action sheet.  You can't tap outside the
  sheet to dismiss it.

The Fix:
  Looks like the view was animated upwards when the keyboard appeared when
the user touched the username (or password) field.  But, when the action sheet
comes up we wern't dismissing the keyboard, nor were we triggering the
animation that moves the main view downwards.  That was causing the action
sheet view to be outside the clipping bounds of its parent container.  Hence,
we could not select anything.

+ When you select an item in the segment controller let's have the text fields
resignFirstResponder (by by keyboard) and also let's trigger the animation that
puts the parent view back where it belongs.  Now you can pick a stage if you
had first touched the un or pw text fields.